feat: fix bugs for deployment
