+++ src/types/feedback.types.ts
@@ -8,6 +8,25 @@ export interface Feedback {
   updatedAt: string;
   type: string;
 }
+export interface Restaurant {
+  id: number;
+  logo: string;
+  slug: string;
+  name: string;
+  favorite: boolean;
+  menus_count: number;
+  available_menus_count: number;
+  menus_shared_count: number;
+  available_menus_shared_count: number;
+  timezone: string;
+  open: boolean;
+  today: null | any;
+  lazy_load_products_recommended: boolean;
+  available_products_count: number;
+  valid_service: boolean;
+  maximum: null | any;
+  header: null | any;
+}
 
 export interface ButtonProps {
   version?: "primary" | "secondary" | "full" | "normal";
